update routers so that delete and put requests now require authentication

+++ blog/controllers/blogs.js
@@ -3,17 +3,6 @@ const Blog = require('../models/blog')
 const logger = require('../utils/logger')
 const User = require('../models/user')
 const { userExtractor } = require('../utils/middleware')
-// const jwt = require('jsonwebtoken')
-
-
-// const getTokenFrom = request => {
-//   const authorization = request.get('authorization')
-//   if (authorization && authorization.startsWith('Bearer ')) {
-//     return authorization.replace('Bearer ', '')
-//   }
-//   return null
-// }
-
 
 blogsRouter.get('/', async (request, response) => {
   logger.info('getting all blogs')
@@ -34,15 +23,13 @@ blogsRouter.get('/:id',  async (request, response) => {
 
 blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
-  // const user = await User.findById(body.userId)
-  const userExtractor = request.user
-  const user = await User.findById(userExtractor.id)
+  const user = await User.findById(request.user.id)
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes,
+    likes: body.likes || 0,
     user: user.id
   })
   if (!blog.likes) {
@@ -55,41 +42,43 @@ blogsRouter.post('/', userExtractor, async (request, response) => {
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
   response.status(201).json(savedBlog)
-  // response.json(savedBlog)
 })
 
 blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = await User.findById(request.user.id)
   const blog = await Blog.findById(request.params.id)
-  // const user = jwt.verify(request.token, process.env.SECRET)
-  const user = request.user
 
-  // console.log('blog',blog)
-  // console.log('user',user)
-  // console.log('bloguser',blog.user)
-  // console.log('bloguser',user.id)
-  if ( blog.user.toString() === user.id.toString() ) {
+  if (blog && blog.user.toString() === user.id.toString()) {
     await Blog.findByIdAndRemove(request.params.id)
-    console.log('bloguser',blog.user.toString())
-    console.log('userid',user.id.toString())
     response.status(204).end()
   } else {
     response.status(401).json({ error: 'Unauthorized' })
   }
 })
 
-
-
 blogsRouter.put('/:id', userExtractor, async (request, response) => {
   const body = request.body
+  const user = await User.findById(request.user.id)
 
-  const blog = {
+  const blog = await Blog.findById(request.params.id)
+
+  if (!blog) {
+    return response.status(404).json({ error: 'Blog not found' })
+  }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'Unauthorized' })
+  }
+
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, {
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes
-  }
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+  }, { new: true })
+
   response.status(202).json(updatedBlog)
 })
+
 
 module.exports = blogsRouter
